fix font issue by removing the "hack"
I thought this was a good idea, but it's definitely not. I made the local font work over the webfont but I neglected a scenario where the user does not have the local font. The webfont woudn't actually work when you do this, so all this time, people has been seeing a Times New Roman font. *facepalm*